requirements.txt
pytest==8.3.2
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.0b2
allure-pytest==2.13.5
allure-python-commons==2.13.5
selenium==4.20.0
pytest-xdist==3.6.1
playwright==1.46.0